Separate useInsertTextSet
